% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.dd.single.R
\name{bp.dd.single}
\alias{bp.dd.single}
\title{Creates control and manipulated Beta-Poisson samples for pre-specified parameters}
\usage{
bp.dd.single(pars,N1,N2,case,degree,seedex)
}
\arguments{
\item{pars}{vector consisting of pre-specified values for the BP parameters alpha, beta, lambda1, lambda2 and p0 (in this order!)}

\item{N1}{size of the sample drawn from the control BP model}

\item{N2}{size of the sample drawn from the manipulated BP model}

\item{case}{case corresponding to a specific parameter manipulation in the BP model; specifically, case can be one of "DLambda", "DAlpha", "DBeta", "DAlphaBeta" or "DPZ" following the design and the respective descriptive table in Schefzik (2020)}

\item{degree}{parameter to set the degree of the created difference (low to strong), see Schefzik (2020) for details and the choice of a range of possible values}

\item{seedex}{seed used for sampling from the fitted BP model to ensure reproducibility}
}
\value{
A list of two:
\itemize{
\item sim.bp.ctrl: vector of simulated control BP sample
\item sim.bp.manip: vector of simulated manipulated BP sample
}
}
\description{
Creates control and manipulated Beta-Poisson (BP) samples for pre-specified BP parameters
}
\details{
Creates control and manipulated BP samples for pre-specified BP parameters
}
\examples{
#create vector consisting of pre-specified values for the BP parameters alpha, beta, lambda1,
#lambda2 and p0 (in this order!)
pars<-c(0.21,3.50,96.13,0.05218,0.02)
N1<-500
N2<-500
degree<-1/3
seedex<-24
dd1<-bp.dd.single(pars,N1,N2,case="DLambda",degree,seedex)
dd2<-bp.dd.single(pars,N1,N2,case="DAlpha",degree,seedex)
dd3<-bp.dd.single(pars,N1,N2,case="DBeta",degree,seedex)
dd4<-bp.dd.single(pars,N1,N2,case="DAlphaBeta",degree,seedex)
dd5<-bp.dd.single(pars,N1,N2,case="DPZ",degree,seedex)


}
\references{
R. Schefzik (2020). Simulating differential distributions in Beta-Poisson models, in particular for single-cell RNA sequencing data.
}
