% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.EX.R
\docType{data}
\name{DATA.EX}
\alias{DATA.EX}
\title{Examplary single-cell RNA-sequencing data set}
\format{
A data matrix containing normalized expression values for 20 genes (rows) and 301 cells (samples; columns)
}
\source{
Excerpt of the data set by Pollen et al. (2014), available at \url{https://hemberg-lab.github.io/scRNA.seq.datasets/human/tissues/}
}
\usage{
data(DATA.EX)
}
\description{
An examplary single-cell RNA-sequencing data set based on an excerpt of the data set by Pollen et al. (2014).
}
\references{
A. A. Pollen, T. J. Nowakowski, J. Shuga, X. Wang, A. A. Leyrat, J. H. Lui, N. Li, L. Szpankowski,B. Fowler, P. Chen, N. Ramalingam, G. Sun, M. Thu, M. Norris, R. Lebofsky, D. Toppani, D. W. Kemp II, M. Wong, B. Clerkson, B. N. Jones, S. Wu, L. Knutsson, B. Alvarado, J. Wang, L. S. Weaver, A. P. May, R. C. Jones, M. A. Unger, A. R. Kriegstein, and J. A. A. West (2014). Low-coverage single-cell mRNA sequencing reveals cellular heterogeneity and activated signaling pathways in developing cerebral cortex. Nature Biotechnology, 32:1053-1058.
}
\keyword{datasets}
